+++ packages/terra-content-container/tests/jest/ContentContainer.test.jsx
@@ -22,6 +22,18 @@ it('should render a header', () => {
   expect(wrapper).toMatchSnapshot();
 });
 
+it('should render a footer', () => {
+  const container = <ContentContainer footer={<p>Footer</p>} />;
+  const wrapper = shallow(container);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render header & footer', () => {
+  const container = <ContentContainer header={<p>Header</p>} footer={<p>Footer</p>} />;
+  const wrapper = shallow(container);
+  expect(wrapper).toMatchSnapshot();
+});
+
 it('should render content', () => {
   const container = (
     <ContentContainer>