+++ packages/terra-site/src/examples/button/Index.jsx
@@ -9,14 +9,12 @@ import { version } from 'terra-button/package.json';
 import ButtonSrc from '!raw-loader!terra-button/src/Button';
 
 // Example Files
-import ButtonSize from './ButtonSize';
 import ButtonBlock from './ButtonBlock';
 import ButtonAnchor from './ButtonAnchor';
 import ButtonDisabled from './ButtonDisabled';
 import ButtonOnClick from './ButtonOnClick';
 import ButtonIcon from './ButtonIcon';
 import ButtonVariant from './ButtonVariant';
-import ButtonCompact from './ButtonCompact';
 
 const ButtonExamples = () => (
   <div>
@@ -25,10 +23,6 @@ const ButtonExamples = () => (
     <PropsTable id="props" src={ButtonSrc} />
     <h2 id="variant">Variant</h2>
     <ButtonVariant />
-    <h2 id="compact">Compact</h2>
-    <ButtonCompact />
-    <h2 id="size">Size</h2>
-    <ButtonSize />
     <h2 id="block">Block</h2>
     <ButtonBlock />
     <h2 id="anchor">Anchor</h2>